+++ database/migrations/2017_10_13_005151_create_seo_table.php
@@ -20,7 +20,7 @@ class CreateSeoTable extends Migration {
                 $table->text('seo_keywords')->nullable();
                 $table->integer('seo_id_connect')->nullable();
                 $table->string('seo_url_connect')->nullable();
-                $table->string('seo_type_connect');
+                $table->string('seo_type_connect')->nullable();
                 $table->timestamps();
 
                 $table->index(['seo_id_connect', 'seo_url_connect']);