+++ views/admin/admin-builder/additional-rows-th.blade.php
@@ -1,5 +1,5 @@
 @if(isset($allowUpdate))
-    <th width="90" class="uk-visible@s"></th>
+    <th width="90" class="@if(isset($app->rows['title'])) uk-visible@s @endif"></th>
 @endif
 @if(isset($allowDestroy))
     <th width="90" class="uk-visible@s"></th>